pyproject.toml
[project]
name = "podcastgpt"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10"
dependencies = [
    "elevenlabs>=2.34.0",
    "fastapi>=0.128.4",
    "google-genai>=1.62.0",
    "motor>=3.7.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pydub>=0.25.1",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "uvicorn[standard]>=0.40.0",
]
